fixed calculation

+++ imageFilteringCpu.cpp
@@ -12,7 +12,7 @@ void imageFilteringCpu
     for(int y = border_size; y < (dst.rows - border_size); y++){
         uchar* pdst = dst.ptr<uchar>(y);
         for(int x = border_size; x < (dst.cols - border_size); x++){
-            double sum = 0.0;
+            float sum = 0.0f;
             for(int yy = 0; yy < kernel.rows; yy++){
                 for(int xx = 0; xx < kernel.cols; xx++){
                     sum += (kernel.ptr<float>(yy)[xx] * src.ptr<uchar>(y+yy-border_size)[x+xx-border_size]);
